Add DSL block style for initializing a call or put option
Many Ruby objects offer the ability to initialize an instance object
using a block. It's a very natural Ruby idiom. Added appropriate
methods to support this idiom and removed the attr_accessors.

Aliased the new methods to the old methods so the original API
would continue to work. Would like to deprecate this old API.

Aliased shorter names for the 'calc' methods. The original method
naming is a bit clunky so this cleans it up.